requirements.txt
asgiref==3.6.0
astroid==3.1.0
dill==0.3.8
Django==4.2.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
dnspython==2.6.1
exceptiongroup==1.2.0
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pathlib2==2.3.7.post1
pathspec==0.5.5
platformdirs==4.2.0
plotly==5.20.0
pluggy==1.4.0
psycopg2==2.9.9
Pygments==2.2.0
pylint==3.1.0
pymongo==4.6.2
pytest==8.1.1
python-dateutil==2.9.0.post0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tabulate==0.9.0
tenacity==8.2.3
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.9.0
tzdata==2024.1
XlsxWriter==3.2.0
zstandard==0.22.0
python-dotenv~=1.0.1